[#922] Rename backup account flow
To differentiate between the account recovery flow and the set backup email
flow, we renamed all resources and url to reflect this. with @anikarni
