+++ web-ui/src/backup_account/backup_account.html
@@ -3,10 +3,10 @@
   <head>
     <meta charset="utf-8"/>
     <meta name="viewport" content="user-scalable=no, width=device-width, initial-scale=1.0, maximum-scale=1.0"/>
-    <title>Pixelated Account Recovery</title>
+    <title>Pixelated Backup Account</title>
   </head>
   <body>
     <div id="root"/>
-    <script type="text/javascript" src="/assets/account_recovery.js"></script>
+    <script type="text/javascript" src="/assets/backup_account.js"></script>
   </body>
 </html>
